Rename file that only exports one function
