+++ src/lib/filter/filter.test.ts
@@ -5,7 +5,7 @@ import { expect, describe, it } from "@jest/globals";
 import { kind, modelToGroupVersionKind } from "kubernetes-fluent-client";
 import * as fc from "fast-check";
 import { CreatePod, DeletePod } from "../../fixtures/loader";
-import { shouldSkipRequest } from "./filter";
+import { shouldSkipRequest } from "./shouldSkipRequest";
 import { AdmissionRequest, Binding, Event } from "../types";
 
 export const callback = () => undefined;
