+++ src/lib/validate-processor.ts
@@ -4,7 +4,7 @@
 import { kind } from "kubernetes-fluent-client";
 
 import { Capability } from "./capability";
-import { shouldSkipRequest } from "./filter/filter";
+import { shouldSkipRequest } from "./filter/shouldSkipRequest";
 import { ValidateResponse } from "./k8s";
 import { AdmissionRequest } from "./types";
 import Log from "./logger";
